+++ src/ui/tileseteditor.cpp
@@ -1067,3 +1067,14 @@ void TilesetEditor::countTileUsage() {
         }
     }
 }
+
+void TilesetEditor::on_copyButton_metatileLabel_clicked() {
+    QClipboard * clipboard = QGuiApplication::clipboard();
+    QString label = this->ui->lineEdit_metatileLabel->text();
+    if (!label.isEmpty()) {
+        Tileset * tileset = Tileset::getMetatileTileset(this->getSelectedMetatileId(), this->primaryTileset, this->secondaryTileset);
+        if (tileset)
+            label.prepend("METATILE_" + QString(tileset->name).replace("gTileset_", "") + "_");
+    }
+    clipboard->setText(label);
+}