+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 - Add Copy/Paste for metatiles in the Tileset Editor.
 - Add ability to set the opacity of the scripting overlay.
 - Add ability to get/set map header properties and read tile pixel data via the API.
+- Add button to copy the full metatile label to the clipboard in the Tileset Editor.
 
 ### Changed
 - If an object event is inanimate, it will always render using its first frame.